+++ proxybroker/cli.py
@@ -30,8 +30,8 @@ def create_parser():
     fparser = subparsers.add_parser(
         'find',
         add_help=False,
-        help='Find and check proxies.',
-        description='Find and check proxies with specified parameters.')
+        help='Find and check proxies',
+        description='Find and check proxies with specified parameters')
     fparser_group = fparser.add_argument_group(title='Options')
     add_find_args(fparser_group)
     add_grab_args(fparser_group)
@@ -43,8 +43,8 @@ def create_parser():
     gparser = subparsers.add_parser(
         'grab',
         add_help=False,
-        help='Find proxies without a check.',
-        description='Find proxies without a check with specified parameters.')
+        help='Find proxies without a check',
+        description='Find proxies without a check with specified parameters')
     gparser_group = gparser.add_argument_group(title='Options')
     add_grab_args(gparser_group)
     add_limit_arg(gparser_group)
@@ -55,18 +55,18 @@ def create_parser():
     sparser = subparsers.add_parser(
         'serve',
         add_help=False,
-        help='Run a local proxy server.',
+        help='Run a local proxy server',
         description='''Run a local proxy server that distributes requests to
                        external proxies, which will be found on the
-                       specified parameters.''')
+                       specified parameters''')
     add_serve_args(sparser.add_argument_group(title='Server options'))
     sparser_fgroup = sparser.add_argument_group(title='Find proxies options')
     add_find_args(sparser_fgroup)
     add_grab_args(sparser_fgroup)
     add_limit_arg(sparser_fgroup, 100, '''
-        When will be found a specified number of working proxies,
-        check of new proxies will be paused.
-        See the documentation for more information.''')
+        When will be found a requested number of working proxies,
+        checking of new proxies will be lazily paused.
+        See the documentation for more information''')
     add_help_arg(sparser.add_argument_group(title='Common options'))
 
     return parser
@@ -78,39 +78,39 @@ def add_broker_args(group):
         type=int,
         default=200,
         dest='max_conn',
-        help='''Limit for the maximum number of concurrent
-                connections when check a proxy.''')
+        help='The maximum number of concurrent checks of proxies')
     group.add_argument(
         '--max-tries',
         type=int,
         default=3,
         dest='max_tries',
-        help='Limit for the maximum number of attempts to check a proxy.')
+        help='The maximum number of attempts to check a proxy')
     group.add_argument(
         '--timeout', '-t',
         type=int,
         default=8,
         metavar='SECONDS',
-        help='''The timeout of a request in seconds.
-                The default value is 8 seconds.''')
+        help='''Timeout of a request in seconds.
+                The default value is 8 seconds''')
     group.add_argument(
         '--judges',
         nargs='+',
-        help='Urls of pages that show HTTP headers and IP address.')
+        help='Urls of pages that show HTTP headers and IP address')
     group.add_argument(
         '--providers',
         nargs='+',
-        help='Urls of pages where to find proxies.')
+        help='Urls of pages where to find proxies')
     group.add_argument(
         '--verify-ssl', '-ssl',
         dest='verify_ssl',
         action='store_true',
-        help='Flag indicating whether to check the SSL certificates.')
+        help='Flag indicating whether to check the SSL certificates')
     group.add_argument(
         '--log',
         nargs='?',
+        default=logging.CRITICAL,
         choices=['NOTSET', 'DEBUG', 'INFO', 'WARNING', 'ERROR', 'CRITICAL'],
-        help='Logging level.')
+        help='Logging level')
     group.add_argument(
         '--version', '-v',
         action='version',
@@ -125,40 +125,40 @@ def add_find_args(group):
         required=True,
         choices=['HTTP', 'HTTPS', 'SOCKS4', 'SOCKS5',
                  'CONNECT:80', 'CONNECT:25'],
-        help='Types (protocols) that need to be check on support by proxy.')
+        help='Type(s) (protocols) that need to be check on support by proxy')
+    group.add_argument(
+        '--lvl',
+        dest='anon_lvl',
+        nargs='+',
+        choices=['Transparent', 'Anonymous', 'High'],
+        help='Level(s) of anonymity (for HTTP only). By default, any level')
     group.add_argument(
         '--data',
         type=argparse.FileType('r'),
         help='''Path to the file with proxies.
-                If specified, used instead of providers.''')
-    group.add_argument(
-        '--anon-lvl',
-        dest='anon_lvl',
-        nargs='+',
-        choices=['Transparent', 'Anonymous', 'High'],
-        help='Level of anonymity (for HTTP only).')
+                If specified, used instead of providers''')
     group.add_argument(
         '--dnsbl',
         nargs='+',
-        help='Spam databases for proxy checking.')
+        help='Spam databases for proxy checking')
     group.add_argument(
         '--post',
         action='store_true',
-        help='Flag indicating use POST instead of GET'
-             'for requests when checking proxies.')
+        help='''Flag indicating use POST instead of GET
+                for requests when checking proxies''')
     group.add_argument(
         '--strict', '-s',
         action='store_true',
         help='''Flag indicating that anonymity levels of the
-                types (protocols) supported by the proxy must
-                be equal to the required types and levels of anonymity.''')
+                types (protocols) supported by a proxy must
+                be equal to the requested types and levels of anonymity''')
 
 
 def add_grab_args(group):
     group.add_argument(
         '--countries', '-c',
         nargs='+',
-        help='Check proxy only from specified countries.')
+        help='List of ISO country codes where should be located proxies')
 
 
 def add_serve_args(group):
@@ -166,33 +166,40 @@ def add_serve_args(group):
         '--host',
         type=str,
         default='127.0.0.1',
-        help='Host of local proxy server.')
+        help='Host of local proxy server')
     group.add_argument(
         '--port',
         type=int,
         default=8888,
-        help='Port of local proxy server.')
+        help='Port of local proxy server')
+    group.add_argument(
+        '--max-tries',
+        type=int,
+        dest='srv_max_tries',
+        help='''The maximum number of attempts to handle an incoming request.
+                If not specified, will be used the value passed to the %(prog)s
+                command''')
     group.add_argument(
         '--min-req-proxy',
         type=int,
         default=5,
         dest='min_req_proxy',
-        help='''Minimum number of processed requests to decide
-                whether to use it further or reject.''')
+        help='''The minimum number of processed requests to decide
+                whether to use it further or reject''')
     group.add_argument(
         '--max-error-rate',
         type=float,
         default=0.5,
         dest='max_error_rate',
-        help='''Maximum percentage of requests that ended
-                with an error. By example: 0.5 = 50%%''')
+        help='''The maximum percentage of requests that ended
+                with an error. For example: 0.5 = 50%%''')
     group.add_argument(
         '--max-resp-time',
         type=int,
         default=8,
         dest='max_resp_time',
         metavar='SECONDS',
-        help='''Maximum response time in seconds. If proxy.avg_resp_time exceeds
+        help='''The maximum response time in seconds. If proxy.avg_resp_time exceeds
                 this value, proxy will be rejected.
                 The default value is 8 seconds''')
     group.add_argument(
@@ -200,20 +207,21 @@ def add_serve_args(group):
         action='store_true',
         dest='prefer_connect',
         help='''Flag that indicates whether to use
-                the CONNECT method if possible.''')
+                the CONNECT method if possible''')
     group.add_argument(
         '--http-allowed-codes',
         nargs='+',
         dest='http_allowed_codes',
-        help='Acceptable HTTP codes returned by proxy on requests.')
+        help='Acceptable HTTP codes returned by proxy on requests')
     group.add_argument(
         '--backlog',
         type=int,
         default=100,
-        help='Maximum number of queued connections passed to listen.')
+        help='The maximum number of queued connections passed to listen')
 
 
-def add_limit_arg(group, _def=0, _help='Maximum number of working proxies.'):
+def add_limit_arg(group, _def=0,
+                  _help='The maximum number of working proxies'):
     group.add_argument(
         '--limit', '-l',
         type=int,
@@ -226,7 +234,7 @@ def add_outfile_arg(group):
         '--outfile', '-o',
         type=argparse.FileType('w'),
         default=sys.stdout,
-        help='Save found proxies to file. By default, output to console.')
+        help='Save found proxies to file. By default, output to console')
 
 
 def add_show_stats_arg(group):
@@ -234,30 +242,18 @@ def add_show_stats_arg(group):
         '--show-stats',
         dest='show_stats',
         action='store_true',
-        help='Flag indicating whether to print verbose stats.')
+        help='Flag indicating whether to print verbose stats')
 
 
 def add_help_arg(group):
     group.add_argument(
         '--help', '-h',
         action='help',
-        help='Show this help message and exit.')
-
-
-def set_logging(level=None):
-    if level:
-        logging.basicConfig(
-            format='%(asctime)s - %(levelname)s - %(name)s - %(message)s',
-            datefmt='[%H:%M:%S]', level=getattr(logging, level))
-
-
-def update_types(ns):
-    if hasattr(ns, 'anon_lvl') and 'HTTP' in ns.types:
-        ns.types.remove('HTTP')
-        ns.types.append(('HTTP', ns.anon_lvl))
+        help='Show this help message and exit')
 
 
 async def handle(proxies, outfile):
+    # TODO: add custom format
     while True:
         proxy = await proxies.get()
         if proxy is None:
@@ -273,8 +269,13 @@ def cli(args=sys.argv[1:]):
         parser.print_help()
         return
 
-    set_logging(level=ns.log)
-    update_types(ns=ns)
+    logging.basicConfig(
+        format='%(asctime)s - %(levelname)s - %(name)s - %(message)s',
+        datefmt='[%H:%M:%S]', level=ns.log)
+
+    if hasattr(ns, 'anon_lvl') and 'HTTP' in ns.types:
+        ns.types.remove('HTTP')
+        ns.types.append(('HTTP', ns.anon_lvl))
 
     loop = asyncio.get_event_loop()
     proxies = asyncio.Queue(loop=loop)
@@ -302,6 +303,7 @@ def cli(args=sys.argv[1:]):
             http_allowed_codes=ns.http_allowed_codes, backlog=ns.backlog,
             data=ns.data, types=ns.types, countries=ns.countries, post=ns.post,
             strict=ns.strict, dnsbl=ns.dnsbl)
+        print('Server started at http://%s:%d' % (ns.host, ns.port))
 
     try:
         if tasks: