+++ CHANGELOG.rst
@@ -1,9 +1,33 @@
+
 Change Log
 ==========
 
 
-`0.1.5`_ (Unreleased)
+`0.2.0-dev`_ (Unreleased)
 ---------------------
+
+* Added CLI interface
+* Added :meth:`Broker.serve` function.
+  Now ProxyBroker can work as a proxy server that distributes incoming requests to a pool of found proxies
+* To available types (protocols) added:
+    * ``CONNECT:80`` - CONNECT method to port 80
+    * ``CONNECT:25`` - CONNECT method to port 25 (SMTP)
+* Added new options of checking and filtering proxies.
+  :meth:`Broker.find` method has takes new parameters:
+  :attr:`post`, :attr:`strict`, :attr:`dnsbl`.
+  See documentation for more information
+* Added check proxies to support Cookies and Referer
+* Added gzip and deflate support
+* :class:`Broker` attributes :attr:`max_concurrent_conn` and :attr:`attempts_conn`
+  are deprecated, use :attr:`max_conn` and :attr:`max_tries` instead.
+* Parameter :attr:`full` in :meth:`Broker.show_stats` is deprecated, use :attr:`verbose` instead
+* Parameter :attr:`types` in :meth:`Broker.find` (and :meth:`Broker.serve`) from now is required
+* :class:`ProxyChecker` renamed to :class:`Checker`.
+  :class:`ProxyChecker` class is deprecated, use :class:`Checker` instead
+* :attr:`Proxy.avgRespTime` renamed to :attr:`Proxy.avg_resp_time`.
+  :attr:`Proxy.avgRespTime` is deprecated, use :attr:`Proxy.avg_resp_time` instead
+* Improved documentation
+* Major refactoring
 
 
 `0.1.4`_ (2016-04-07)
@@ -16,7 +40,7 @@ Change Log
 ---------------------
 
 * ``ProxyProvider`` renamed to ``Provider``.
-  ``ProxyProvider`` class is deprecated, use ``Provider`` instead
+  ``ProxyProvider`` class is deprecated, use ``Provider`` instead.
 * ``Broker`` now accepts a list of providers and judges not only as strings 
   but also objects of classes ``Provider`` and ``Judge``
 * Fixed bug with signal handler on Windows `#4`_
@@ -89,4 +113,4 @@ Change Log
 .. _0.1.2: https://github.com/constverum/ProxyBroker/compare/v0.1...v0.1.2
 .. _0.1.3: https://github.com/constverum/ProxyBroker/compare/v0.1.2...v0.1.3
 .. _0.1.4: https://github.com/constverum/ProxyBroker/compare/v0.1.3...v0.1.4
-.. _0.1.5: https://github.com/constverum/ProxyBroker/compare/v0.1.4...HEAD
+.. _0.2.0-dev: https://github.com/constverum/ProxyBroker/compare/v0.1.4...HEAD